Split square.test.ts (2096→1112 lines) into two files
Extracts createMockClient helper to src/test-utils/square-helpers.ts so both
split files can share it without duplication. The first file keeps connection,
client lifecycle, createPaymentLink, retrieveOrder, and retrievePayment tests.
The second file (square-payment-provider.test.ts, 972 lines) covers refund,
webhook verification, REST transport, and provider integration tests.
